+++ src/daf/core.py
@@ -23,7 +23,6 @@ __all__ = (
     "shutdown",
     "add_object",
     "remove_object",
-    "get_guild_user",
     "get_accounts",
     "initialize"
 )
@@ -58,15 +57,8 @@ async def cleanup_accounts():
 
 
 @misc.doc_category("DAF control reference")
-async def initialize(token: Optional[str] = None,
-                     server_list: Optional[List[Union[guild.GUILD, guild.USER, guild.AutoGUILD]]] = None,
-                     is_user: Optional[bool] = False,
-                     user_callback: Optional[Union[Callable, Coroutine]] = None,
-                     server_log_output: Optional[str] = None,
-                     sql_manager: Optional[sql.LoggerSQL] = None,
-                     intents: Optional[discord.Intents] = None,
-                     debug: Optional[Union[TraceLEVELS, int, str, bool]] = TraceLEVELS.NORMAL,
-                     proxy: Optional[str] = None,
+async def initialize(user_callback: Optional[Union[Callable, Coroutine]] = None,
+                     debug: Optional[Union[TraceLEVELS, int, str]] = TraceLEVELS.NORMAL,
                      logger: Optional[logging.LoggerBASE] = None,
                      accounts: Optional[List[client.ACCOUNT]] = []) -> None:
     """
@@ -84,10 +76,6 @@ async def initialize(token: Optional[str] = None,
     # ------------------------------------------------------------
     # Initialize tracing
     # ------------------------------------------------------------
-    if isinstance(debug, bool):
-        trace("Using bool for debug parameter is DEPRECATED. Use daf.logging.TraceLEVELS", TraceLEVELS.DEPRECATED)
-        debug = TraceLEVELS.NORMAL if debug else TraceLEVELS.DEPRECATED
-
     if debug is None:
         debug = TraceLEVELS.NORMAL
 
@@ -95,24 +83,6 @@ async def initialize(token: Optional[str] = None,
     # ------------------------------------------------------------
     # Initialize logging
     # ------------------------------------------------------------
-    # --------------- DEPRECATED -------------------- #
-    if server_log_output is not None:
-        if logger is None:
-            trace("Using 'server_log_output' is planned for removal.\n"
-                  "Use logger=LoggerJSON(path=\"{server_log_output}\") instead.",
-                  TraceLEVELS.DEPRECATED)
-            logger = logging.LoggerJSON(path=server_log_output)
-        else:
-            trace("'logger' parameter was passed, ignoring server_log_output", TraceLEVELS.WARNING)
-
-    if sql_manager is not None:
-        if logger is None:
-            trace("'sql_manager' is scheduled for removal\nUse logger=LoggerSQL(...) instead.",
-                  TraceLEVELS.DEPRECATED)
-            logger = sql_manager
-        else:
-            trace("logger parameter was passed, ignoring sql_manager", TraceLEVELS.WARNING)
-    # ----------------------------------------------- #
     if logger is None:
         logger = logging.LoggerJSON(path="History")
 
@@ -122,14 +92,6 @@ async def initialize(token: Optional[str] = None,
     # ------------------------------------------------------------
     # Initialize accounts
     # ------------------------------------------------------------
-    # ------------- DEPRECATED -----------------
-    if token is not None:
-        trace("Passing the token argument directly is deprecated since v2.4 where support\n"
-              "for multiple accounts was added. Please use the ``accounts`` parameter",
-              TraceLEVELS.DEPRECATED)
-
-        accounts.append(client.ACCOUNT(token=token, is_user=is_user, intents=intents, proxy=proxy, servers=server_list))
-    # ------------------------------------------
     for account in accounts:
         try:
             await add_object(account)
@@ -197,9 +159,6 @@ async def add_object(obj: Union[guild.USER, guild.GUILD, guild.AutoGUILD],
         The object provided is not supported for addition.
     TypeError
         Invalid parameter type.
-    RuntimeError
-        When using deprecated method of adding items to the shill list,
-        no accounts were available.
     Other
         Raised in the obj.initialize() method
     """
@@ -211,11 +170,6 @@ async def add_object(obj: Union[guild.USER, guild.GUILD, guild.AutoGUILD],
 async def add_object(obj: Union[message.DirectMESSAGE, message.TextMESSAGE, message.VoiceMESSAGE],
                      snowflake: Union[guild.GUILD, guild.USER]) -> None:
     """
-    .. deprecated:: v2.4
-
-        Using int, discord.* objects in the snowflake parameter.
-        This functionality is planned for removal in v2.5.
-
     Adds a message to the daf.
 
     Parameters
@@ -253,16 +207,6 @@ async def add_object(obj, snowflake=None):
         await obj.initialize()
         GLOBALS.accounts.append(obj)
     elif isinstance(obj, (guild._BaseGUILD, guild.AutoGUILD)):
-        if snowflake is None:
-            # Compatibility with prior versions of v2.4
-            trace("Directly adding guild like objects to the framework is deprecated since v2.4 (multi-account support)\n"
-                  "The object will be added to the first account in the list. Update your code to pass ``snowflake`` with :class:`~daf.client.ACCOUNT`",
-                  TraceLEVELS.DEPRECATED)
-            try:
-                snowflake = GLOBALS.accounts[0]
-            except IndexError as exc:
-                raise RuntimeError("No accounts are running in the framework") from exc
-
         if not isinstance(snowflake, client.ACCOUNT):
             raise TypeError("snowflake parameter type must be ACCOUNT when the obj parameter type is guild like.")
 
@@ -273,16 +217,7 @@ async def add_object(obj, snowflake=None):
             raise ValueError("snowflake parameter (guild-like) is required to add a message.")
 
         if not isinstance(snowflake, (guild.AutoGUILD, guild._BaseGUILD)):
-            # --------- DEPRECATED ----------- #
-            # TODO: remove in v2.5, uncomment TypError
-            snowflake = get_guild_user(snowflake)
-            trace("Using int or discord.* objects is deprecated for the snowflake parameter of add_object.\n"
-                  "It is planned for removal in version v2.5!",
-                  TraceLEVELS.DEPRECATED)
-            if snowflake is None:
-                raise ValueError("The GUILD/USER with specified snowflake could not be found.")
-            # -------------------------------- #
-            # raise TypeError("snowflake parameter must be one of: guild.AutoGUILD, guild.GUILD, guild.USER")
+            raise TypeError("snowflake parameter must be one of: guild.AutoGUILD, guild.GUILD, guild.USER")
 
         await snowflake.add_message(obj)
 
@@ -333,45 +268,6 @@ async def remove_object(
     elif isinstance(snowflake, client.ACCOUNT):
         await snowflake._close()
         GLOBALS.accounts.remove(snowflake)
-
-
-@typechecked
-def get_guild_user(
-    snowflake: Union[int, discord.Object, discord.Guild, discord.User, discord.Object]
-) -> Union[guild.GUILD, guild.USER, None]:
-    """
-    TODO: Remove in v2.5
-    .. deprecated:: v2.4
-
-    Retrieves the GUILD/USER object that has the ``snowflake`` ID from the shilling list.
-
-    Parameters
-    -------------
-    snowflake: Union[int, discord.Object, discord.Guild, discord.User, discord.Object]
-        Snowflake ID or discord objects containing snowflake id of the GUILD.
-
-    Raises
-    ---------------
-    TypeError
-        Incorrect snowflake type
-
-    Returns
-    ---------------
-    :class:`daf.guild.GUILD` | :class:`daf.guild.USER`
-        The object requested.
-    None
-        If not guild/user not in the shilling list.
-    """
-
-    trace("This function is planned for removal in the future. Use ACCOUNT.get_server method instead!", TraceLEVELS.DEPRECATED)
-    if isinstance(snowflake, int):
-        snowflake = discord.Object(snowflake)
-
-    for guild in GLOBALS.accounts[0]._servers:
-        if guild.snowflake == snowflake.id:
-            return guild
-
-    return None
 
 
 @misc.doc_category("Clients")
@@ -438,50 +334,27 @@ def _shutdown_clean(loop: asyncio.AbstractEventLoop) -> None:
 
 @typechecked
 @misc.doc_category("DAF control reference")
-def run(token: Optional[str] = None,
-        server_list: Optional[List[Union[guild.GUILD, guild.USER, guild.AutoGUILD]]] = None,
-        is_user: Optional[bool] = False,
-        user_callback: Optional[Union[Callable, Coroutine]] = None,
-        server_log_output: Optional[str] = None,
-        sql_manager: Optional[sql.LoggerSQL] = None,
-        intents: Optional[discord.Intents] = None,
+def run(user_callback: Optional[Union[Callable, Coroutine]] = None,
         debug: Optional[Union[TraceLEVELS, int, str, bool]] = TraceLEVELS.NORMAL,
-        proxy: Optional[str] = None,
         logger: Optional[logging.LoggerBASE] = None,
         accounts: Optional[List[client.ACCOUNT]] = []) -> None:
     """
+    .. versionchanged:: 2.7
+
+       Removed deprecated parameters (see :ref:`v2.7`)
+
     Runs the framework and does not return until the framework is stopped (:func:`daf.core.shutdown`).
     After stopping, it returns None.
 
-    .. warning::
-        This will block until the framework is stopped, if you want manual control over the
-        asyncio event loop, eg. you want to start the framework as a task, use
-        the :func:`daf.core.initialize` coroutine.
-
-
-    .. versionchanged:: v2.4
-        Added ``accounts`` parameter.
-
-    .. deprecated:: v2.4
-
-        .. card::
-
-            The following parameters were deprecated in favor of support for multiple accounts
-            ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
-            - token
-            - is_user
-            - server_list
-            - intents
-            - proxy
-            +++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++
-            The above parameters should be passed to :class:`~daf.client.ACCOUNT`.
-
+    This will block until the framework is stopped, if you want manual control over the
+    asyncio event loop, eg. you want to start the framework as a task, use
+    the :func:`daf.core.initialize` coroutine.
 
     Parameters
     ---------------
     user_callback: Optional[Union[Callable, Coroutine]]
         Function or async function to call after the framework has been started.
-    debug : Optional[TraceLEVELS | int | str] = TraceLEVELS.NORMAL
+    debug: Optional[TraceLEVELS | int | str] = TraceLEVELS.NORMAL
         .. versionchanged:: v2.3
             Deprecate use of bool (assume TraceLEVELS.NORMAL).
             Add support for TraceLEVELS or int or str that converts to TraceLEVELS.